+++ packages/react-aria/src/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from './useAriaButton';
+export * from './useARIAButton';
